fields.h: properly size the field
field was wrongly typed as int, but in reality it should store a size_t.

This fixes compilation issues on platforms where sizeof(int) != sizeof(size_t).

+++ code/game/fields.h
@@ -66,7 +66,7 @@ typedef enum
 typedef struct
 {
 	char	*psName;
-	int		iOffset;
+	size_t		iOffset;
 	fieldtypeSAVE_t	eFieldType;
 	int		iFlags;
 } field_t;
